change extract user_id

+++ models/claim.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Claims struct {
-	UserID   int32  `json:"user_id"`
+	UserID   uint64 `json:"user_id"`
 	Username string `json:"username"`
 	jwt.StandardClaims
 }
