+++ queries/models.go
@@ -9,7 +9,7 @@ import (
 )
 
 type User struct {
-	ID        int32     `json:"id"`
+	ID        uint64     `json:"id"`
 	Name      string    `json:"name"`
 	Email     string    `json:"email"`
 	Photo     string    `json:"photo"`
