+++ queries/user.sql.go
@@ -97,7 +97,7 @@ SELECT id, name, email, photo, verified, password, role, created_at, updated_at 
 WHERE id = $1 LIMIT 1
 `
 
-func (q *Queries) GetUserById(ctx context.Context, id int32) (User, error) {
+func (q *Queries) GetUserById(ctx context.Context, id uint64) (User, error) {
 	row := q.db.QueryRowContext(ctx, getUserById, id)
 	var i User
 	err := row.Scan(
